feat(nav)!: remove the top bar, add the mobile bottom bar
The sidebar now owns all global chrome on desktop. Below 1024px a
three-slot fixed bar carries Menu, Search and Items; account moves into
the drawer footer drop-up.

Everything the top bar hosted has a home already: search moved to an
overlay (Cmd/Ctrl-K plus a sidebar row), and language, theme and account
moved into the sidebar footer drop-up.

Shell grid loses its top row at every breakpoint, switches to 100dvh,
and the content area gains bottom padding so the fixed bar never covers
the last row of a table. The bar itself pads for env(safe-area-inset-bottom).

Deletes top_bar.rs, apps_menu.rs (AppsMenu had been unreferenced dead
code; UserMenu is superseded by AccountMenu) and LanguageNavMenu.
