+++ ultros-frontend/ultros-app/src/components/mod.rs
@@ -7,7 +7,6 @@ pub mod add_to_list;
 pub mod alert_config_drawer;
 pub mod alert_rules_panel;
 pub mod app_shell;
-pub mod apps_menu;
 pub mod chart_toolbar;
 pub mod cheapest_price;
 pub mod clipboard;
@@ -42,6 +41,7 @@ pub mod market_heat;
 pub mod market_movers;
 pub mod market_pulse;
 pub mod meta;
+pub mod mobile_bar;
 pub mod modal;
 pub mod number_input;
 pub mod on_hand_input;
@@ -74,7 +74,6 @@ pub mod toggle;
 pub mod tool_help;
 pub mod toolbar;
 pub mod tooltip;
-pub mod top_bar;
 pub mod top_opportunity;
 pub mod ui_text;
 pub mod undercut_alert_drawer;
